Basic (working) entry screen
Here is a basic entry screen with a working New Game button and placeholders for the Resume Game and Load Game buttons. It also includes the font image files for improvements I'll add to the game title on the entry screen.
Please note: I made a couple of changes to Entity.swift (lines 118 and 124) to get the code to compile, but I'm not sure that these changes completely solve the issues, so please review those changes.
More to come tonight!